fix(predicate): Break out boolean logic from trait
BREAKING CHANGE: `Predicate` trait is changed but most use cases should
not be impacted.
