+++ src/core.rs
@@ -6,7 +6,6 @@
 // option. This file may not be copied, modified, or distributed
 // except according to those terms.
 
-use boolean::{AndPredicate, NotPredicate, OrPredicate};
 use boxed::BoxPredicate;
 
 /// Trait for generically evaluating a type against a dynamically created
@@ -20,64 +19,6 @@ pub trait Predicate<Item: ?Sized> {
     /// Execute this `Predicate` against `variable`, returning the resulting
     /// boolean.
     fn eval(&self, variable: &Item) -> bool;
-
-    /// Compute the logical AND of two `Predicate` results, returning the result.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use predicates::prelude::*;
-    ///
-    /// let predicate_fn1 = predicate::always().and(predicate::always());
-    /// let predicate_fn2 = predicate::always().and(predicate::never());
-    /// assert_eq!(true, predicate_fn1.eval(&4));
-    /// assert_eq!(false, predicate_fn2.eval(&4));
-    fn and<B>(self, other: B) -> AndPredicate<Self, B, Item>
-    where
-        B: Predicate<Item>,
-        Self: Sized,
-    {
-        AndPredicate::new(self, other)
-    }
-
-    /// Compute the logical OR of two `Predicate` results, returning the result.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use predicates::prelude::*;
-    ///
-    /// let predicate_fn1 = predicate::always().or(predicate::always());
-    /// let predicate_fn2 = predicate::always().or(predicate::never());
-    /// let predicate_fn3 = predicate::never().or(predicate::never());
-    /// assert_eq!(true, predicate_fn1.eval(&4));
-    /// assert_eq!(true, predicate_fn2.eval(&4));
-    /// assert_eq!(false, predicate_fn3.eval(&4));
-    fn or<B>(self, other: B) -> OrPredicate<Self, B, Item>
-    where
-        B: Predicate<Item>,
-        Self: Sized,
-    {
-        OrPredicate::new(self, other)
-    }
-
-    /// Compute the logical NOT of a `Predicate`, returning the result.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use predicates::prelude::*;
-    ///
-    /// let predicate_fn1 = predicate::always().not();
-    /// let predicate_fn2 = predicate::never().not();
-    /// assert_eq!(false, predicate_fn1.eval(&4));
-    /// assert_eq!(true, predicate_fn2.eval(&4));
-    fn not(self) -> NotPredicate<Self, Item>
-    where
-        Self: Sized,
-    {
-        NotPredicate::new(self)
-    }
 
     /// Returns a `BoxPredicate` wrapper around this `Predicate` type.
     ///
